read seller products

+++ routes/web.php
@@ -5,6 +5,8 @@ use App\Http\Controllers\Web\Backend\Auth\AuthController;
 use App\Http\Controllers\Web\Backend\Root\Dashboard\DashboardController;
 use App\Http\Controllers\Web\Backend\Root\ProductCategory\ProductCategoryController as AppProductCategoryController;
 use App\Http\Controllers\Web\Backend\Root\ProductSubCategory\ProductSubCategoryController as AppProductSubCategoryController;
+use App\Http\Controllers\Web\Backend\Seller\Dashboard\DashboardController as AppDashboardController;
+use App\Http\Controllers\Web\Backend\Seller\Product\ProductController as AppProductController;
 use App\Http\Controllers\Web\Frontend\Cart\CartController;
 use App\Http\Controllers\Web\Frontend\Home\HomeController;
 use App\Http\Controllers\Web\Frontend\Product\ProductController;
@@ -56,26 +58,36 @@ Route::middleware(['auth:sanctum', config('jetstream.auth_session'), 'verified']
     Route::get('/check-point', [CheckPointController::class, 'checkPoint'])->name('check-point');
     // Route for administrator
     Route::prefix('administrator')->middleware('role:Developer|Administrator')->group(function () {
-        Route::get('dashboard', [DashboardController::class, 'index'])->name('admin.dashboards');
+        Route::get('dashboard', [DashboardController::class, 'index'])->name('admin.dashboards')->middleware('can:read-admin-dashboards');
         // Product Category Route
-        Route::prefix('product-categories')->middleware('can:read-product-categories')->group(function () {
-            Route::get('', [AppProductCategoryController::class, 'index'])->name('admin.product-categories')->middleware('can:read-product-categories');
-            Route::get('get-data', [AppProductCategoryController::class, 'getData'])->name('admin.product-categories.get-data')->middleware('can:read-product-categories');
-            Route::post('store', [AppProductCategoryController::class, 'store'])->name('admin.product-categories.store')->middleware('can:create-product-categories');
-            Route::get('{productCategory}/show', [AppProductCategoryController::class, 'show'])->name('admin.product-categories.update')->middleware('can:update-product-categories');
-            Route::post('{productCategory}/update', [AppProductCategoryController::class, 'update'])->name('admin.product-categories.update')->middleware('can:update-product-categories');
-            Route::delete('{productCategory}/delete', [AppProductCategoryController::class, 'destroy'])->name('admin.product-categories.delete')->middleware('can:delete-product-categories');
-            Route::get('{productCategory}/update-status', [AppProductCategoryController::class, 'updateStatus'])->name('admin.product-categories.update-status')->middleware('can:update-product-categories');
+        Route::prefix('product-categories')->middleware('can:read-admin-product-categories')->group(function () {
+            Route::get('', [AppProductCategoryController::class, 'index'])->name('admin.product-categories')->middleware('can:read-admin-product-categories');
+            Route::get('get-data', [AppProductCategoryController::class, 'getData'])->name('admin.product-categories.get-data')->middleware('can:read-admin-product-categories');
+            Route::post('store', [AppProductCategoryController::class, 'store'])->name('admin.product-categories.store')->middleware('can:create-admin-product-categories');
+            Route::get('{productCategory}/show', [AppProductCategoryController::class, 'show'])->name('admin.product-categories.update')->middleware('can:update-admin-product-categories');
+            Route::post('{productCategory}/update', [AppProductCategoryController::class, 'update'])->name('admin.product-categories.update')->middleware('can:update-admin-product-categories');
+            Route::delete('{productCategory}/delete', [AppProductCategoryController::class, 'destroy'])->name('admin.product-categories.delete')->middleware('can:delete-admin-product-categories');
+            Route::get('{productCategory}/update-status', [AppProductCategoryController::class, 'updateStatus'])->name('admin.product-categories.update-status')->middleware('can:update-admin-product-categories');
         });
         // Product Sub Category Route
-        Route::prefix('product-sub-categories')->middleware('can:read-product-sub-categories')->group(function () {
-            Route::get('', [AppProductSubCategoryController::class, 'index'])->name('admin.product-sub-categories')->middleware('can:read-product-sub-categories');
-            Route::get('get-data', [AppProductSubCategoryController::class, 'getData'])->name('admin.product-sub-categories.get-data')->middleware('can:read-product-sub-categories');
-            Route::post('store', [AppProductSubCategoryController::class, 'store'])->name('admin.product-sub-categories.store')->middleware('can:create-product-sub-categories');
-            Route::get('{productSubCategory}/show', [AppProductSubCategoryController::class, 'show'])->name('admin.product-sub-categories.update')->middleware('can:update-product-sub-categories');
-            Route::post('{productSubCategory}/update', [AppProductSubCategoryController::class, 'update'])->name('admin.product-sub-categories.update')->middleware('can:update-product-sub-categories');
-            Route::delete('{productSubCategory}/delete', [AppProductSubCategoryController::class, 'destroy'])->name('admin.product-sub-categories.delete')->middleware('can:delete-product-sub-categories');
-            Route::get('{productSubCategory}/update-status', [AppProductSubCategoryController::class, 'updateStatus'])->name('admin.product-sub-categories.update-status')->middleware('can:update-product-sub-categories');
+        Route::prefix('product-sub-categories')->middleware('can:read-admin-product-sub-categories')->group(function () {
+            Route::get('', [AppProductSubCategoryController::class, 'index'])->name('admin.product-sub-categories')->middleware('can:read-admin-product-sub-categories');
+            Route::get('get-data', [AppProductSubCategoryController::class, 'getData'])->name('admin.product-sub-categories.get-data')->middleware('can:read-admin-product-sub-categories');
+            Route::post('store', [AppProductSubCategoryController::class, 'store'])->name('admin.product-sub-categories.store')->middleware('can:create-admin-product-sub-categories');
+            Route::get('{productSubCategory}/show', [AppProductSubCategoryController::class, 'show'])->name('admin.product-sub-categories.update')->middleware('can:update-admin-product-sub-categories');
+            Route::post('{productSubCategory}/update', [AppProductSubCategoryController::class, 'update'])->name('admin.product-sub-categories.update')->middleware('can:update-admin-product-sub-categories');
+            Route::delete('{productSubCategory}/delete', [AppProductSubCategoryController::class, 'destroy'])->name('admin.product-sub-categories.delete')->middleware('can:delete-admin-product-sub-categories');
+            Route::get('{productSubCategory}/update-status', [AppProductSubCategoryController::class, 'updateStatus'])->name('admin.product-sub-categories.update-status')->middleware('can:update-admin-product-sub-categories');
+        });
+    });
+
+    Route::prefix('seller')->middleware('role:Seller')->group(function () {
+        Route::get('dashboard', [AppDashboardController::class, 'index'])->name('seller.dashboards');
+
+        // Product
+        Route::prefix('products')->middleware('can:read-seller-products')->group(function () {
+            Route::get('', [AppProductController::class, 'index'])->name('seller.products');
+            Route::get('get-data', [AppProductController::class, 'getData'])->name('seller.products.get-data');
         });
     });
 });
